Optimize dashboard reorder speed (#74)
Dashboard reordering is now a much more efficient operation. This was
achieved by maintaining a separate list of items that we are editing,
rather than waiting for the database to save data and tell us about it.

Note there is currently no means to save the modified list, but having a
superior UX is a more important step here. UI components for other bits
coming soon.